corrected materials to avoid openmc segfault

+++ tests/test_neutronics_model.py
@@ -18,11 +18,12 @@ class TestShape(unittest.TestCase):
 
         local_filename = 'dagmc_bigger.h5m'
 
-        r = requests.get(url, stream=True)
-        with open(local_filename, 'wb') as f:
-            for chunk in r.iter_content(chunk_size=1024): 
-                if chunk:
-                    f.write(chunk)
+        if not Path(local_filename).is_file():
+            r = requests.get(url, stream=True)
+            with open(local_filename, 'wb') as f:
+                for chunk in r.iter_content(chunk_size=1024): 
+                    if chunk:
+                        f.write(chunk)
                     
         self.h5m_filename_bigger = local_filename
 
@@ -36,13 +37,15 @@ class TestShape(unittest.TestCase):
         url = "https://github.com/fusion-energy/neutronics_workflow/raw/main/example_01_single_volume_cell_tally/stage_2_output/dagmc.h5m"
 
         local_filename = 'dagmc_smaller.h5m'
-        # NOTE the stream=True parameter
-        r = requests.get(url, stream=True)
-        with open(local_filename, 'wb') as f:
-            for chunk in r.iter_content(chunk_size=1024): 
-                if chunk: # filter out keep-alive new chunks
-                    f.write(chunk)
-                    #f.flush() commented by recommendation from J.F.Sebastian
+        
+
+        if not Path(local_filename).is_file():
+            r = requests.get(url, stream=True)
+            with open(local_filename, 'wb') as f:
+                for chunk in r.iter_content(chunk_size=1024): 
+                    if chunk: # filter out keep-alive new chunks
+                        f.write(chunk)
+
         self.h5m_filename_smaller = local_filename
